commit working energy code

+++ main.cu
@@ -33,16 +33,16 @@ int filePrompt() {
         fgets(line, MAX_LINE_LENGTH, file);
 
         int i = 0;
-        float temp_mass;
-        float temp_radius;
-        float3 temp_position;
-        float3 temp_velocity;
+        double temp_mass;
+        double temp_radius;
+        double3 temp_position;
+        double3 temp_velocity;
 
         while(fgets(line, MAX_LINE_LENGTH, file) != NULL && i < 10) {
             printf(line);
             printf("\n");
             
-            sscanf(line, "%f,%f,%f,%f,%f,%f,%f,%f", 
+            sscanf(line, "%lf,%lf,%lf,%lf,%lf,%lf,%lf,%lf", 
                 &temp_mass,
                 &temp_radius,
                 &temp_position.x,
@@ -53,16 +53,16 @@ int filePrompt() {
                 &temp_velocity.z);
             
             bi[i].id = i;
-            bi[i].mass = (float) temp_mass;
-            bi[i].radius = (float) temp_radius;
+            bi[i].mass = (double) temp_mass;
+            bi[i].radius = (double) temp_radius;
             
-            bi[i].position.x = (float) temp_position.x;
-            bi[i].position.y = (float) temp_position.y;
-            bi[i].position.z = (float) temp_position.z;
+            bi[i].position.x = (double) temp_position.x;
+            bi[i].position.y = (double) temp_position.y;
+            bi[i].position.z = (double) temp_position.z;
 
-            bi[i].velocity.x = (float) temp_velocity.x;
-            bi[i].velocity.y = (float) temp_velocity.y;
-            bi[i].velocity.z = (float) temp_velocity.z;
+            bi[i].velocity.x = (double) temp_velocity.x;
+            bi[i].velocity.y = (double) temp_velocity.y;
+            bi[i].velocity.z = (double) temp_velocity.z;
 
             i++;
         }
@@ -137,15 +137,15 @@ int main (int argc, char *argv[]) {
 
     unsigned long long max_ticks = timerPrompt(); 
 
-    float secs_per_tick = tickTime(); //1 by default
+    double secs_per_tick = tickTime(); //1 by default
 
     /* auto scaling code */
     struct body origin;
     origin.position.x = 0;
     origin.position.y = 0;
     origin.position.z = 0;
-    float max_distance = 0;
-    float autoscale;
+    double max_distance = 0;
+    double autoscale;
 
     double totalmass;
     double totalenergy_k;
